+++ lib/read/cardano-wallet-read.cabal
@@ -55,6 +55,7 @@ library
     Cardano.Read.Ledger.Tx.Eras
     Cardano.Read.Ledger.Tx.ExtraSigs
     Cardano.Read.Ledger.Tx.Fee
+    Cardano.Read.Ledger.Tx.Hash
     Cardano.Wallet.Read
     Cardano.Wallet.Read.Block
     Cardano.Wallet.Read.Block.BHeader
@@ -86,7 +87,6 @@ library
     Cardano.Wallet.Read.Tx.Gen.Mary
     Cardano.Wallet.Read.Tx.Gen.Shelley
     Cardano.Wallet.Read.Tx.Gen.TxParameters
-    Cardano.Wallet.Read.Tx.Hash
     Cardano.Wallet.Read.Tx.Inputs
     Cardano.Wallet.Read.Tx.Integrity
     Cardano.Wallet.Read.Tx.Metadata
