+++ lib/wallet/src/Cardano/Wallet/DB/Store/Transactions/TransactionInfo.hs
@@ -27,6 +27,9 @@ import Cardano.Read.Ledger.Tx.CollateralOutputs
 import Cardano.Read.Ledger.Tx.Fee
     ( getEraFee
     )
+import Cardano.Read.Ledger.Tx.Hash
+    ( getEraTxHash
+    )
 import Cardano.Read.Ledger.Tx.Inputs
     ( getEraInputs
     )
@@ -103,9 +106,6 @@ import Cardano.Wallet.Read.Eras
     ( EraValue
     , IsEra
     , applyEraFun
-    )
-import Cardano.Wallet.Read.Tx.Hash
-    ( getEraTxHash
     )
 import Cardano.Wallet.Read.Tx.Metadata
     ( getEraMetadata
